fix hover on rune showing tooltip, fix often restarting when maging flat resistances, fix custom script with bug showing 'script correct' icon + without reset button, remove digits dataset from main ocr (it was slow), crocoring script no more vitality overmage, test for % res + range exos (not passing yet)

+++ Inkybot/Services/ExoItemMageResolve.cs
@@ -56,6 +56,6 @@ namespace Inkybot.Services
         }
         
         protected override int Priority(ItemMage itemMage) =>
-            (int) itemMage.Rune.Sink - (itemMage.Rune.Sink <= Sink ? 1 : 0) * 1000;
+            (int) itemMage.Rune.Sink - ((int)itemMage.Rune.Sink <= Sink ? 1 : 0) * 1000;
     }
 }
